# static/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: NW Monitor API
  description: |
    REST API for the Network Monitor System.
    Provides endpoints for device management, monitoring, topology,
    dashboards, alerts, SLA reports, bandwidth, SNMP traps, syslog,
    user management, audit logs, and system administration.
  version: "1.2.0"
  contact:
    name: NW Monitor Admin
servers:
  - url: /
    description: Current server
...
paths:
...
  /api/devices:
    get:
      tags: [Devices]
      summary: List all devices
      responses:
        "200":
          description: Array of devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
    post:
      tags: [Devices]
      summary: Add a new device
      description: Requires operator role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceCreate"
      responses:
        "201":
          description: Device created
        "400":
          description: Validation error
  /api/devices/{device_id}:
    put:
      tags: [Devices]
      summary: Update a device
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceCreate"
      responses:
        "200":
          description: Device updated
    delete:
      tags: [Devices]
      summary: Delete a device
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      responses:
        "200":
          description: Device deleted
...
  /api/devices/import/csv:
    post:
      tags: [Devices]
      summary: Import devices from CSV
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Import results

  # ---- Monitoring ----
...
  /api/statistics/trend:
    get:
      tags: [Monitoring]
      summary: Get statistics trend data
      parameters:
        - name: hours
          in: query
          schema: { type: integer, default: 24 }
      responses:
        "200":
          description: Trend data
  /api/check/{device_id}:
    post:
      tags: [Monitoring]
      summary: Trigger an immediate check for a device
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      responses:
        "200":
          description: Check result

  # ---- SNMP ----
  /api/snmp/{device_id}/interfaces:
    get:
      tags: [SNMP]
      summary: Get SNMP interface data
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      responses:
        "200":
          description: Interface list
  /api/snmp/{device_id}/custom-oids:
    get:
      tags: [SNMP]
      summary: Get custom OIDs for a device
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      responses:
        "200":
          description: Custom OID list
    post:
      tags: [SNMP]
      summary: Add a custom OID to a device
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                oid:
                  type: string
                label:
                  type: string
      responses:
        "201":
          description: OID added
  /api/snmp/custom-oids/{oid_id}:
    delete:
      tags: [SNMP]
      summary: Delete a custom OID
      parameters:
        - name: oid_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OID deleted
  /api/snmp/{device_id}/custom-oids/query:
    post:
      tags: [SNMP]
      summary: Query custom OIDs for a device
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      responses:
        "200":
          description: OID query results

  # ---- Topology ----
...
components:
  parameters:
    DeviceId:
      name: device_id
      in: path
      required: true
      schema: { type: integer }
...
  schemas:
    Device:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        ip_address: { type: string }
        device_type: { type: string }
        location: { type: string }
        monitor_type: { type: string, enum: [ping, http, snmp, tcp, dns] }
        status: { type: string, enum: [up, down, unknown] }
        response_time: { type: number }
        last_check: { type: string, format: date-time }
    DeviceCreate:
      type: object
      required: [name, ip_address]
      properties:
        name: { type: string }
        ip_address: { type: string }
        device_type: { type: string }
        location: { type: string }
        monitor_type: { type: string, enum: [ping, http, snmp, tcp, dns], default: ping }
        snmp_community: { type: string }
        snmp_port: { type: integer, default: 161 }
        snmp_version: { type: string, enum: ["1", "2c", "3"] }
        tcp_port: { type: integer }
        dns_query_domain: { type: string }
        location_type: { type: string, enum: [on-premise, cloud] }
...